requirements.txt
# Core
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# AI/ML
langchain>=0.0.350
langchain-community>=0.0.10
ollama>=0.1.7
sentence-transformers>=2.2.2
huggingface-hub>=0.16.4
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1

# Utilities
python-dateutil>=2.8.2
pydantic>=2.5.0
email-validator>=2.1.0
requests>=2.31.0
tqdm>=4.65.0  # For progress bars

# Development
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
